Creating a user data validations class
