+++ src/main/java/com/agilethought/atsceapi/exception/ErrorMessage.java
@@ -12,5 +12,4 @@ public class ErrorMessage {
     public static final String BAD_REQUEST_MESSAGE_EMAIL_IN_DB = "can not be store, because it already exists in the Database, try with another email";
     public static final String BAD_REQUEST_MESSAGE_STATUS = "The status needs to be 0 or 1";
     public static final String BAD_REQUEST_MESSAGE_UNMATCHED_EMAIL = "Email provided does not match";
-    public static final String BAD_REQUEST_MESSAGE_INCORRECT_UPPERCASE = "Is not valid the lowercase";
 }
