+++ src/main/java/com/agilethought/atsceapi/validator/UpdateValidator.java
@@ -11,7 +11,6 @@ import org.springframework.stereotype.Service;
 import java.util.Optional;
 
 import static com.agilethought.atsceapi.exception.ErrorMessage.*;
-import static com.agilethought.atsceapi.validator.ValidationUtils.*;
 
 @Service
 public class UpdateValidator implements Validator<UpdateUserRequest> {
@@ -19,15 +18,13 @@ public class UpdateValidator implements Validator<UpdateUserRequest> {
     @Autowired
     private UserRepository userRepository;
 
+    @Autowired
+    private UserDataValidationUtils userDataValidationUtils;
+
     @Override
     public void validate(UpdateUserRequest userUp) {
+        userDataValidationUtils.validateUserData(userUp.getType(),userUp.getFirstName(),userUp.getLastName(),userUp.getEmail(),userUp.getPassword(),userUp.getStatus());
         validateUserId(userUp.getId());
-        validateUserTypeField(userUp.getType());
-        validateUserFirstName(userUp.getFirstName());
-        validateUserLastName(userUp.getLastName());
-        validateEmailFormat(userUp.getEmail());
-        validatePasswordFormat(userUp.getPassword());
-        validateUserStatus(userUp.getStatus());
         validateEmailInTheDataBase(userUp.getEmail(), userUp.getId());
     }
 
@@ -41,45 +38,6 @@ public class UpdateValidator implements Validator<UpdateUserRequest> {
         Optional<User> user = userRepository.findById(id);
         if (user.isPresent() && !user.get().getEmail().equals(email) && userRepository.existsByEmail(email)){
             throw new BadRequestException("The *EMAIL*" + " " + BAD_REQUEST_MESSAGE_EMAIL_IN_DB);
-        }
-    }
-
-    private void validateUserStatus(Integer status) {
-        if(status != null) {
-            if( status < 0 || status > 1) {
-                throw new BadRequestException(BAD_REQUEST_MESSAGE_STATUS);
-            }
-        }
-    }
-
-    private void validatePasswordFormat(String password) {
-
-        if(!isStringValid(password) || !isStringLowerCase(password) || !isValidPassword(password)) {
-            throw new BadRequestException(BAD_REQUEST_MESSAGE + "*PASSWORD*" + " " + BAD_REQUEST_MESSAGE_PASSWORD);
-        }
-    }
-
-    private void validateUserLastName(String lastName) {
-        if(!isStringValid(lastName)) {
-            throw new BadRequestException(BAD_REQUEST_MESSAGE + "*LASTNAME*" + " " + BAD_REQUEST_MESSAGE_FIRST_NAME_LASTNAME);
-        }
-    }
-
-    private void validateEmailFormat(String email) {
-        if(!isStringValid(email) || !isStringLowerCase(email) || !isValidEmail(email)) {
-            throw new BadRequestException(BAD_REQUEST_MESSAGE + "*EMAIL*" + " " + BAD_REQUEST_MESSAGE_EMAIL);
-        }
-    }
-
-    private void validateUserFirstName(String firstName) {
-        if(!isStringValid(firstName)) {
-            throw new BadRequestException(BAD_REQUEST_MESSAGE + "*NAME*" + " " + BAD_REQUEST_MESSAGE_FIRST_NAME_LASTNAME);
-        }
-    }
-
-    private void validateUserTypeField(int type) {
-        if(type < 1 || type > 2) {
-            throw new BadRequestException(BAD_REQUEST_MESSAGE + "*TYPE*" + " " + BAD_REQUEST_MESSAGE_TYPE);
         }
     }
