+++ src/main/resources/application.properties
@@ -1,2 +1,2 @@
-#spring.data.mongodb.uri=${MONGO_URI}
+spring.data.mongodb.uri=${MONGO_URI}
 server.error.include-message=always
